InGameScreen을 화면 방향 반응형으로 변경
