+++ app/src/main/java/com/example/sharon/screens/Start.kt
@@ -1,0 +1,75 @@
+package com.example.sharon.screens
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.geometry.Size
+import androidx.compose.ui.layout.onGloballyPositioned
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import androidx.compose.ui.unit.toSize
+
+class Start {
+    companion object {
+        @Composable
+        fun StartScreen(nextScreen: () -> Unit) {
+            var containerSize by remember { mutableStateOf(Size.Zero) }
+
+            Scaffold { innerPadding ->
+                Box(
+                    modifier = Modifier
+                        .background(MaterialTheme.colorScheme.background)
+                        .fillMaxSize()
+                        .padding(innerPadding)
+                ) {
+                    Column(
+                        verticalArrangement = Arrangement.Center,
+                        horizontalAlignment = Alignment.CenterHorizontally,
+                        modifier = Modifier
+                            .padding(16.dp)
+                            .fillMaxSize()
+                            .onGloballyPositioned { coordinates ->
+                                containerSize = coordinates.size.toSize()
+                            }
+                    ) {
+                        Text(
+                            text = "무궁화 꽃이",
+                            fontSize = (containerSize.width * 7/100).toInt().sp,
+                            color = MaterialTheme.colorScheme.primary
+                        )
+                        Text(
+                            text = "피었습니다!",
+                            fontSize = (containerSize.width * 7/100).toInt().sp,
+                            color = MaterialTheme.colorScheme.primary
+                        )
+                        Button(
+                            onClick = {nextScreen()},
+                            colors = ButtonDefaults.buttonColors(
+                                containerColor = MaterialTheme.colorScheme.primary,
+                                contentColor = MaterialTheme.colorScheme.onPrimary
+                            ),
+                            modifier = Modifier.align(Alignment.CenterHorizontally)
+                        ) {
+                            Text("다음 화면")
+                        }
+                    }
+                }
+            }
+        }
+    }
+}
